% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtx.R
\name{gtx}
\alias{gtx}
\title{Create a *gtx* table object}
\usage{
gtx(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{...}{Accepts all arguments in \code{as_workbook()}.}
}
\value{
Extended \code{gt} object with additional workbook object configured based on \code{gt} options
}
\description{
Create a *gtx* table object
}
\examples{

dplyr::starwars |> gtx()

}
