% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_workbook.R
\name{as_workbook}
\alias{as_workbook}
\title{Convert to Excel workbook object}
\usage{
as_workbook(.data, .sheet_name = NULL, .start_col = 2, .start_row = 2)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{.sheet_name}{Which worksheet to write to. It can be the worksheet index or name. Default is \code{NULL}.}

\item{.start_col}{A counting number specifying the starting column to write to.}

\item{.start_row}{A counting number specifying the starting row to write to.}
}
\value{
Workbook object with \code{gt} options applied.
}
\description{
Convert to Excel workbook object
}
\examples{

dplyr::starwars |>
  gtx() |>
  as_workbook()
}
